requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
scikit-image==0.21.0
Flask==2.3.2
Flask-CORS==4.0.0
pytest==7.4.0
jupyter==1.0.0
pandas==2.0.3
tqdm==4.65.0
